# [SPARK-45452][SQL] Improve `InMemoryFileIndex` to use `FileSystem.listFiles` API