### What changes were proposed in this pull request?

This PR aims to use `InMemoryFileIndex` to use `FileSystem.listFiles` API instead of many recursive FileSystem calls.

To do this, a new configuration is introduced with the default value, `s3a`.

```
spark.sql.sources.useListFilesFileSystemList=s3a
```

### Why are the changes needed?

To improve the nested directory listing performance. Currently, it's designed to apply for a single root directory listing case on S3A file system. For example, the improvement on a 3-year-data table with `year/month/day/hour` hierarchy is **158x**.

**3-YEAR PARTITIONED TABLE (2021 ~ 2023)**
```
$ bin/spark-sql
spark-sql (default)> CREATE TABLE t USING CSV PARTITIONED BY (year, month, day, hour) LOCATION "file:///tmp/three_year" AS SELECT /*+ REPARTITION(64, year, month, day, hour) */ Y.id year, M.id month, D.id day, H.id hour, 0 value FROM RANGE(2021, 2024) Y, RANGE(1, 13) M, RANGE(1, 31) D, RANGE(0, 24) H;

$ aws s3 sync three_year s3://dongjoon/three_year

$ aws s3 ls s3://dongjoon/three_year --summarize --recursive | more
2023-10-07 01:07:00          8 three_year/._SUCCESS.crc
2023-10-07 01:07:00          0 three_year/_SUCCESS
2023-10-07 01:07:00         12 three_year/year=2021/month=1/day=1/hour=0/.part-00033-264e1d55-d466-43c8-8972-283a1f84082b.c000.csv.crc
2023-10-07 01:07:00          2 three_year/year=2021/month=1/day=1/hour=0/part-00033-264e1d55-d466-43c8-8972-283a1f84082b.c000.csv
...

$ aws s3 ls s3://dongjoon/three_year --summarize --recursive | tail -n2
Total Objects: 51842
   Total Size: 362888
```

**AFTER (11s)**
```
$ bin/spark-shell \
-c spark.hadoop.fs.s3a.aws.credentials.provider=com.amazonaws.auth.profile.ProfileCredentialsProvider
...
scala> spark.time(new org.apache.spark.sql.execution.datasources.InMemoryFileIndex(spark, Seq(new org.apache.hadoop.fs.Path("s3a://dongjoon/three_year")), Map.empty, None))
Time taken: 11270 ms
```

**BEFORE (1741s)**
- I used one year to estimate 3 years at this time. I'll update with 3 year result soon.
```
$ bin/spark-shell \
-c spark.hadoop.fs.s3a.aws.credentials.provider=com.amazonaws.auth.profile.ProfileCredentialsProvider \
-c spark.sql.sources.useListFilesFileSystemList=none
...
scala> spark.time(new org.apache.spark.sql.execution.datasources.InMemoryFileIndex(spark, Seq(new org.apache.hadoop.fs.Path("s3a://dongjoon/three_year")), Map.empty, None))
Time taken: 1741315 ms
```

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Pass the CIs with the newly added test case.

### Was this patch authored or co-authored using generative AI tooling?

No.