@dongjoon-hyun 2023-10-07
The last commit is about fixing two typos in the string and comment. And, the previous commit passed all tests. Let me merge this for Apache Spark 4.0.0. Thank you again, @viirya .

![Screenshot 2023-10-07 at 4 51 17 PM](https://github.com/apache/spark/assets/9700541/46f9cb27-10e8-453a-83da-150bfc346da5)