@cloud-fan 2023-12-19
Have we tested it on a flattened directory layout (no partitions but many files)? Does it have perf regression?