@cloud-fan 2023-12-20
Just FYI, we hit a 10x perf regression with this change in our internal benchmark with a flattened directory layout. This PR says it can replace many recursive FileSystem calls, which is not an issue for flattened directories. Shall we disable it by default? Or we only use it when scanning a file source table and the table metadata says it has more than one partition column.