@dongjoon-hyun 2024-01-04
Thank you for sharing and sorry for being late, @cloud-fan .  I was on a vacation from December 14th.

Could you elaborate a little more specific? For example, how many files do you have in a single S3 prefix? What is the exact time value *before* and *after* when you say `10x`?
> Just FYI, we hit a 10x perf regression with this change in our internal benchmark with a flattened directory layout. 

Of course, we can, @cloud-fan . Before doing that, I want to reproduce your benchmark to make it sure and to improve for that case.
> Shall we disable it by default?