@cloud-fan 2024-01-04
The time went from 10 seconds to more than 1 minute. I don't know the exact number of files but it should be a lot as it took 10 seconds to list files.