@cloud-fan 2024-01-04
> As you see in this PR description, the listing time of 51842 files was 11 seconds in my experiment.

Did you try with a flattened directory?