@dongjoon-hyun 2024-01-04
Not yet. I'm trying to understand your case.

BTW, may I ask what do you mean by `a flattened directory`? So far, I thought what you mean is a table on a directory (without any nesting or partitions) with 1M files? Is it some other technique?