@cloud-fan 2024-01-04
> is a table on a directory (without any nesting or partitions) with 1M files?

yea, no partition directories.