@cloud-fan 2024-01-04
@dongjoon-hyun  Thanks for the confirmation! I need to check with the benchmark team and verify if it's caused by extra storage layers (e.g. DBFS) not supporting this new API well.